Move cli commands to its own module
Affinity has a dedicate cli module for its commands. That said move
commands that were on different modules to the cli module.
